discord_bot/discord copy/requirements.txt
discord~=2.3.2
openai~=1.14.0
nextcord~=2.6.0
python-dotenv~=1.0.1
requests~=2.31.0
pillow~=10.2.0